パスワードを隠す機能の追加、checkInfoの削除
