--- lib/view/start_up/sign_in/sign_in_model.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/cupertino.dart';
-
-class SignInModel extends ChangeNotifier {
-  final formKey = GlobalKey<FormState>();
-  TextEditingController emailController = TextEditingController();
-  TextEditingController passwordController = TextEditingController();
-
-  void checkInfo() {
-    if (emailController.text == '') {
-      throw 'メールアドレスが入力されていません';
-    }
-    if (passwordController.text == '') {
-      throw 'パスワードが入力されていません';
-    }
-  }
-}
